+++ src/middleware.ts
@@ -4,7 +4,7 @@ import { cookies } from "next/headers";
 import { i18n } from "./i18n";
 import { match as matchLocale } from "@formatjs/intl-localematcher";
 import Negotiator from "negotiator";
-import { getLanguageCookie } from "./utils/cookies";
+import { auth } from "./auth";
 
 // 로그인 세션 확인 여부 체크
 async function checkLogin() {
@@ -18,12 +18,11 @@ function requiresAuth(pathname: string, urls: string[]) {
   return urls.some((url) => !!match(url)(pathname));
 }
 
-export function getLocale(request: NextRequest): string | undefined {
-  // 쿠키에서 사용자 언어 확인
-  const cookieLocale = getLanguageCookie(request);
-
-  if (cookieLocale) {
-    return cookieLocale;
+export async function getLocale(request: NextRequest): Promise<string | undefined> {
+  // 사용자 세션 확인
+  const userSession = await auth();
+  if (userSession?.user.language && i18n.locales.includes(userSession.user.language)) {
+    return userSession.user.language;
   }
 
   // Negotiator expects plain object so we need to transform headers
@@ -46,7 +45,7 @@ export const middleware = async (req: NextRequest) => {
   const url = req.nextUrl.clone();
 
   // Check if there is any supported locale in the pathname
-  const locale = getLocale(req) || i18n.defaultLocale;
+  const locale = (await getLocale(req)) || i18n.defaultLocale;
   const pathnameIsMissingLocale = i18n.locales.every(
     (locale) => !pathname.startsWith(`/${locale}/`) && pathname !== `/${locale}`,
   );
